Break MarkedVector out of ArgList.h. https://bugs.webkit.org/show_bug.cgi?id=309107 rdar://171662292
Reviewed by Yusuke Suzuki.

MarkedVector is a distinct class.  This is also in preparation for implementation other subclasses of MarkedVector.

No new tests needed.  This is purely a refactoring patch.

* Source/JavaScriptCore/CMakeLists.txt:
* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/Sources.txt:
* Source/JavaScriptCore/runtime/ArgList.cpp:
(JSC::MarkedVectorBase::addMarkSet): Deleted.
(JSC::MarkedVectorBase::markLists): Deleted.
(JSC::MarkedVectorBase::slowEnsureCapacity): Deleted.
(JSC::MarkedVectorBase::expandCapacity): Deleted.
(JSC::MarkedVectorBase::slowAppend): Deleted.
* Source/JavaScriptCore/runtime/ArgList.h:
(JSC::MarkedVector::MarkedVector): Deleted.
(JSC::MarkedVector::begin): Deleted.
(JSC::MarkedVector::end): Deleted.
(JSC::MarkedVector::at const): Deleted.
(JSC::MarkedVector::set): Deleted.
(JSC::MarkedVector::clear): Deleted.
(JSC::MarkedVector::append): Deleted.
(JSC::MarkedVector::appendWithCrashOnOverflow): Deleted.
(JSC::MarkedVector::last const): Deleted.
(JSC::MarkedVector::takeLast): Deleted.
(JSC::MarkedVector::ensureCapacity): Deleted.
(JSC::MarkedVector::hasOverflowed): Deleted.
(JSC::MarkedVector::fill): Deleted.
(JSC::MarkedVector::isUsingInlineBuffer const): Deleted.
* Source/JavaScriptCore/runtime/MarkedVector.cpp: Copied from Source/JavaScriptCore/runtime/ArgList.cpp.
(JSC::MarkedVectorBase::addMarkSet):
(JSC::MarkedVectorBase::markLists):
(JSC::MarkedVectorBase::slowEnsureCapacity):
(JSC::MarkedVectorBase::expandCapacity):
(JSC::MarkedVectorBase::slowAppend):
* Source/JavaScriptCore/runtime/MarkedVector.h: Copied from Source/JavaScriptCore/runtime/ArgList.h.
(JSC::MarkedVector::MarkedVector):
(JSC::MarkedVector::begin):
(JSC::MarkedVector::end):
(JSC::MarkedVector::at const):
(JSC::MarkedVector::set):
(JSC::MarkedVector::clear):
(JSC::MarkedVector::append):
(JSC::MarkedVector::appendWithCrashOnOverflow):
(JSC::MarkedVector::last const):
(JSC::MarkedVector::takeLast):
(JSC::MarkedVector::ensureCapacity):
(JSC::MarkedVector::hasOverflowed):
(JSC::MarkedVector::fill):
(JSC::MarkedVector::isUsingInlineBuffer const):

Canonical link: https://commits.webkit.org/308583@main
