+++ Source/JavaScriptCore/CMakeLists.txt
@@ -1268,6 +1268,7 @@ set(JavaScriptCore_PRIVATE_FRAMEWORK_HEADERS
     runtime/LeafExecutable.h
     runtime/Lookup.h
     runtime/MachineContext.h
+    runtime/MarkedVector.h
     runtime/MatchResult.h
     runtime/MathCommon.h
     runtime/MathObject.h
